use linux types

+++ src/main.c
@@ -16,20 +16,22 @@ void __init my_start_kernel(void)
 	/* Initialize process 0 */
 	task[pid].pid = pid;
 	task[pid].state = 0; /* -1 unrunnable, 0 runnable, >0 stopped */
-	task[pid].task_entry = task[pid].thread.ip = (unsigned long) my_process;
+	task[pid].task_entry = task[pid].thread.ip = (uintptr_t) my_process;
 	task[pid].thread.sp =
-		(unsigned long) &task[pid].stack[KERNEL_STACK_SIZE - 1];
+		(uintptr_t) &task[pid].stack[KERNEL_STACK_SIZE - 1];
 	task[pid].next = &task[pid];
+
 	/* fork more process */
 	for (i = 1; i < MAX_TASK_NUM; i++) {
 		memcpy(&task[i], &task[0], sizeof(tPCB));
 		task[i].pid = i;
 		task[i].state = -1;
 		task[i].thread.sp =
-			(unsigned long) &task[i].stack[KERNEL_STACK_SIZE - 1];
+			(uintptr_t) &task[i].stack[KERNEL_STACK_SIZE - 1];
 		task[i].next = task[i-1].next;
 		task[i-1].next = &task[i];
 	}
+
 	/* start process 0 by task[0] */
 	pid = 0;
 	my_current_task = &task[pid];
